requirements.txt
requests
pyarrow
joblib
black
pandas
boto3
scikit-learn
tensorflow==2.7.0
pyyaml
matplotlib